med-agi/services/imaging/app/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
python-jose[cryptography]==3.3.0
httpx==0.25.1
numpy==1.24.3
pillow==10.1.0
pydicom==2.4.3
tritonclient[all]==2.40.0
prometheus-client==0.19.0
python-multipart==0.0.6